requirements.txt
Flask==1.0.2
flask_environments==0.1
google-cloud-storage==1.13.2
gunicorn==19.9.0
# oauth2client==4.1.3
Flask-SQLAlchemy==2.3.2
PyMySQL==0.9.2
six==1.11.0
oauth2client==4.1.3
flask-cors==3.0.7